ImprovedGamepad Improvement
Mostly a refactor for naming, but some fixes too.

Button -> InputMechanism
BooleanButton -> Button
FloatButton -> AnalogInput
Created a new Joystick class, which does not extend InputMechanism.

Deleted class JoystickButton, and AngleUtilities.

There was some incomplete and/or duplicate ways to get at the same data before. Now this is cleaned up, and you have to get at the joystick data through the Joystick class.

The only thing which uses AngleUtilities is Joystick, and that is pretty minimal, so I just put the two trig functions in there directly.

Other fixes/features:
* Now we do have the stick click/button.
* Added Config decorator for FTCDashboard
* Made dead zones public static so they could be seen changed by Dashboard
* Since we have a lot of good abstraction going on here, and Dashboard is pretty rigid, also checked for Config changes and then pass those into the classes. This works here because gamepad is called in the loop/update.