Fixes for ShapeDetection
Add Config decorator for ShapeDetection, so that it can be changed from dashboard. Change access modifiers so that config-things are public static, as required. Make other things private.

Try to avoid churning OpenCV memory in the pipeline -- create the Mats and such only once. Still seeing a notice that memory is 'leaking' (garbage collected), although I'm not sure where. Could be in the operations like draw contour? Commented out the copymakeborder for this purpose.